+++ isis/tests/FunctionalTestsFootprintinit.cpp
@@ -38,7 +38,7 @@ TEST_F(DefaultCube, FunctionalTestFootprintinitDefault) {
   std::vector<double> lats = {9.928502, 9.928502, 10.434859, 10.434859, 9.928502};
 
   geos::geom::CoordinateArraySequence coordArray = geos::geom::CoordinateArraySequence(*(boundary->getCoordinates()));
-  for (int i = 0; i < coordArray.getSize(); i++) {
+  for (size_t i = 0; i < coordArray.getSize(); i++) {
     EXPECT_NEAR(lons[i], coordArray.getAt(i).x, 1e-6);
     EXPECT_NEAR(lats[i], coordArray.getAt(i).y, 1e-6);
   }
@@ -60,7 +60,7 @@ TEST_F(DefaultCube, FunctionalTestFootprintinitLincSinc) {
   std::vector<double> lats = {9.928500, 9.928500, 10.434861, 10.434861, 9.928500};
 
   geos::geom::CoordinateArraySequence coordArray = geos::geom::CoordinateArraySequence(*(boundary->getCoordinates()));
-  for (int i = 0; i < coordArray.getSize(); i++) {
+  for (size_t i = 0; i < coordArray.getSize(); i++) {
     EXPECT_NEAR(lons[i], coordArray.getAt(i).x, 1e-6);
     EXPECT_NEAR(lats[i], coordArray.getAt(i).y, 1e-6);
   }
@@ -82,7 +82,7 @@ TEST_F(DefaultCube, FunctionalTestFootprintinitVertices) {
   std::vector<double> lats = {9.928456, 9.928456, 10.434903, 10.434903, 9.928456};
 
   geos::geom::CoordinateArraySequence coordArray = geos::geom::CoordinateArraySequence(*(boundary->getCoordinates()));
-  for (int i = 0; i < coordArray.getSize(); i++) {
+  for (size_t i = 0; i < coordArray.getSize(); i++) {
     EXPECT_NEAR(lons[i], coordArray.getAt(i).x, 1e-6);
     EXPECT_NEAR(lats[i], coordArray.getAt(i).y, 1e-6);
   }
@@ -104,7 +104,7 @@ TEST_F(DefaultCube, FunctionalTestFootprintinitCamera) {
   std::vector<double> lats = {9.924583, 9.924583, 10.329275, 10.329275, 9.924583};
 
   geos::geom::CoordinateArraySequence coordArray = geos::geom::CoordinateArraySequence(*(boundary->getCoordinates()));
-  for (int i = 0; i < coordArray.getSize(); i++) {
+  for (size_t i = 0; i < coordArray.getSize(); i++) {
     EXPECT_NEAR(lons[i], coordArray.getAt(i).x, 1e-6);
     EXPECT_NEAR(lats[i], coordArray.getAt(i).y, 1e-6);
   }
@@ -126,7 +126,7 @@ TEST_F(DefaultCube, FunctionalTestFootprintinitTestXY) {
   std::vector<double> lats = {9.928502, 9.928502, 10.434859, 10.434859, 9.928502};
 
   geos::geom::CoordinateArraySequence coordArray = geos::geom::CoordinateArraySequence(*(boundary->getCoordinates()));
-  for (int i = 0; i < coordArray.getSize(); i++) {
+  for (size_t i = 0; i < coordArray.getSize(); i++) {
     EXPECT_NEAR(lons[i], coordArray.getAt(i).x, 1e-6);
     EXPECT_NEAR(lats[i], coordArray.getAt(i).y, 1e-6);
   }
@@ -154,7 +154,7 @@ TEST_F(DefaultCube, FunctionalTestFootprintinitPrecision) {
   std::vector<double> lats = {9.928502, 9.928502, 10.434859, 10.434859, 9.928502};
 
   geos::geom::CoordinateArraySequence coordArray = geos::geom::CoordinateArraySequence(*(boundary->getCoordinates()));
-  for (int i = 0; i < coordArray.getSize(); i++) {
+  for (size_t i = 0; i < coordArray.getSize(); i++) {
     EXPECT_NEAR(lons[i], coordArray.getAt(i).x, 1e-6);
     EXPECT_NEAR(lats[i], coordArray.getAt(i).y, 1e-6);
   }
